Move tokenize to its own crate
This involves removing the series impl of tokenize and changing tests.
