+++ Cargo.toml
@@ -117,7 +117,8 @@ members = [
   "src/daft-scan",
   "src/daft-scheduler",
   "src/daft-sketch",
-  "src/daft-functions"
+  "src/daft-functions",
+  "src/daft-tokenize"
 ]
 
 [workspace.dependencies]
