+++ daft/daft.pyi
@@ -1288,8 +1288,6 @@ class PySeries:
     def utf8_to_date(self, format: str) -> PySeries: ...
     def utf8_to_datetime(self, format: str, timezone: str | None = None) -> PySeries: ...
     def utf8_normalize(self, remove_punct: bool, lowercase: bool, nfd_unicode: bool, white_space: bool) -> PySeries: ...
-    def tokenize_encode(self, tokens_path: str) -> PySeries: ...
-    def tokenize_decode(self, tokens_path: str) -> PySeries: ...
     def is_nan(self) -> PySeries: ...
     def is_inf(self) -> PySeries: ...
     def not_nan(self) -> PySeries: ...
