+++ src/daft-core/src/kernels/mod.rs
@@ -1,4 +1,3 @@
 pub mod hashing;
 pub mod search_sorted;
-pub mod tokenize;
 pub mod utf8;
