+++ src/daft-tokenize/Cargo.toml
@@ -1,0 +1,24 @@
+[dependencies]
+arrow2 = {workspace = true, features = [
+  "chrono-tz",
+  "compute_take",
+  "compute_cast",
+  "compute_aggregate",
+  "compute_if_then_else",
+  "compute_sort",
+  "compute_filter",
+  "compute_temporal",
+  "compute_comparison",
+  "compute_arithmetics",
+  "compute_concatenate",
+  "compute_substring",
+  "io_ipc"
+]}
+common-error = {path = "../common/error", default-features = false}
+daft-core = {path = "../daft-core", default-features = false}
+tiktoken-rs = "0.5.9"
+
+[package]
+name = "daft-tokenize"
+edition.workspace = true
+version.workspace = true
